refactor(experimental): graphql: shuffle around repository layout
